fixes session loading when forum not yet installed
