+++ src/User/UserServiceProvider.php
@@ -14,10 +14,9 @@ namespace Flarum\User;
 use Flarum\Event\ConfigureUserPreferences;
 use Flarum\Event\GetPermission;
 use Flarum\Foundation\AbstractServiceProvider;
+use Flarum\Http\SessionServiceProvider;
 use Illuminate\Contracts\Container\Container;
-use Illuminate\Session\FileSessionHandler;
 use RuntimeException;
-use SessionHandlerInterface;
 
 class UserServiceProvider extends AbstractServiceProvider
 {
@@ -26,22 +25,10 @@ class UserServiceProvider extends AbstractServiceProvider
      */
     public function register()
     {
-        $this->registerSession();
+        $this->app->register(SessionServiceProvider::class);
+
         $this->registerGate();
         $this->registerAvatarsFilesystem();
-    }
-
-    protected function registerSession()
-    {
-        $this->app->singleton('session.handler', function ($app) {
-            return new FileSessionHandler(
-                $app['files'],
-                $app['config']['session.files'],
-                $app['config']['session.lifetime']
-            );
-        });
-
-        $this->app->alias('session.handler', SessionHandlerInterface::class);
     }
 
     protected function registerGate()
